backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Google Sheets integration
gspread==6.0.2
google-auth==2.27.0

# Gemini AI
google-genai>=1.0.0

# HTTP client for Telegram API
httpx==0.26.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0